chore: generate types with `yarn generate-dts`
