+++ types/wc-toast-close-button.d.ts
@@ -1,0 +1,12 @@
+/**
+ * @cssprop [--wc-toast-stroke=#2a2a32] - stroke. Default is #2a2a32;
+ *
+ * @summary - button to dismiss toast item
+ *
+ * @tag wc-toast-close-button
+ */
+export default class WCToastCloseButton extends HTMLElement {
+    static template(): string;
+    template: HTMLTemplateElement;
+}
+//# sourceMappingURL=wc-toast-close-button.d.ts.map
